# apps/demo/internal/docs/swagger.yaml
definitions:
  dto.DefaultRender:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  go-gin-web_apps_demo_internal_dto_dtoUser.UserCreateReq:
    properties:
      companyId:
        description: 公司id
        type: integer
      departmentId:
        description: 部门id
        type: integer
      name:
        description: 姓名
        type: string
    type: object
  go-gin-web_apps_demo_internal_dto_dtoUser.UserCreateResp:
    properties:
      id:
        description: 数据自增id
        type: integer
    type: object
  go-gin-web_apps_demo_internal_dto_dtoUser.UserDeleteReq:
    properties:
      id:
        description: 数据自增id
        type: integer
    required:
    - id
    type: object
  go-gin-web_apps_demo_internal_dto_dtoUser.UserDetailResp:
    properties:
      companyId:
        description: 公司id
        type: integer
      createdAt:
        description: 创建时间
        type: integer
      createdBy:
        description: 创建人id
        type: integer
      departmentId:
        description: 部门id
        type: integer
      id:
        description: 数据自增id
        type: integer
      name:
        description: 姓名
        type: string
      updatedAt:
        description: 更新时间
        type: integer
      updatedBy:
        description: 更新人id
        type: integer
    required:
    - id
    type: object
  go-gin-web_apps_demo_internal_dto_dtoUser.UserPageListItem:
    properties:
      companyId:
        description: 公司id
        type: integer
      createdAt:
        description: 创建时间
        type: integer
      createdBy:
        description: 创建人id
        type: integer
      departmentId:
        description: 部门id
        type: integer
      id:
        description: 数据自增id
        type: integer
      name:
        description: 姓名
        type: string
      updatedAt:
        description: 更新时间
        type: integer
      updatedBy:
        description: 更新人id
        type: integer
    required:
    - id
    type: object
  go-gin-web_apps_demo_internal_dto_dtoUser.UserPageListResp:
    properties:
      list:
        description: 数据列表
        items:
          $ref: '#/definitions/go-gin-web_apps_demo_internal_dto_dtoUser.UserPageListItem'
        type: array
      total:
        description: 数据总条数
        type: integer
    type: object
  go-gin-web_apps_demo_internal_dto_dtoUser.UserUpdateReq:
    properties:
      companyId:
        description: 公司id
        type: integer
      departmentId:
        description: 部门id
        type: integer
      id:
        description: 数据自增id
        type: integer
      name:
        description: 姓名
        type: string
    required:
    - id
    type: object
info:
  contact: {}
paths:
  /demo/user/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建用户
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-gin-web_apps_demo_internal_dto_dtoUser.UserCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0,"data": "ok","msg": "success"}'
          schema:
            allOf:
            - $ref: '#/definitions/dto.DefaultRender'
            - properties:
                data:
                  $ref: '#/definitions/go-gin-web_apps_demo_internal_dto_dtoUser.UserCreateResp'
              type: object
      summary: 创建用户
      tags:
      - 用户管理
  /demo/user/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: 删除用户
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-gin-web_apps_demo_internal_dto_dtoUser.UserDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0,"data": "ok","msg": "删除成功"}'
          schema:
            allOf:
            - $ref: '#/definitions/dto.DefaultRender'
            - properties:
                data:
                  type: string
              type: object
      summary: 删除用户
      tags:
      - 用户管理
  /demo/user/detail:
    get:
      consumes:
      - application/json
      parameters:
      - description: 数据自增id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0,"data": "ok","msg": "success"}'
          schema:
            allOf:
            - $ref: '#/definitions/dto.DefaultRender'
            - properties:
                data:
                  $ref: '#/definitions/go-gin-web_apps_demo_internal_dto_dtoUser.UserDetailResp'
              type: object
      summary: 用户详情
      tags:
      - 用户管理
  /demo/user/pageList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数据条数
        in: query
        maximum: 1000
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0,"data": "ok","msg": "success"}'
          schema:
            allOf:
            - $ref: '#/definitions/dto.DefaultRender'
            - properties:
                data:
                  $ref: '#/definitions/go-gin-web_apps_demo_internal_dto_dtoUser.UserPageListResp'
              type: object
      summary: 用户列表分页
      tags:
      - 用户管理
  /demo/user/update:
    post:
      consumes:
      - application/json
      parameters:
      - description: 修改用户
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-gin-web_apps_demo_internal_dto_dtoUser.UserUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0,"data": "ok","msg": "修改成功"}'
          schema:
            allOf:
            - $ref: '#/definitions/dto.DefaultRender'
            - properties:
                data:
                  type: string
              type: object
      summary: 修改用户
      tags:
      - 用户管理
swagger: "2.0"